refactor: Update statistics routing
