+++ src/components/mypage/statistics/ShortAnswerStatistics.tsx
@@ -1,4 +1,4 @@
-import StatisticsInfoHeader from '@/components/mypage/StatisticsInfoHeader';
+import StatisticsInfoHeader from '@/components/mypage/statistics/StatisticsInfoHeader';
 
 interface ShortAnswerStatisticsProps {
   title: string;
